Shift tag package to be inside person package for src/test, rename predicate test files
